requirements.txt
numpy
tensorflow-cpu
matplotlib
opencv-python
imageio